requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
idna==3.4
imageio==2.22.4
munch==2.5.0
networkx==2.8.8
numpy==1.23.5
opencv-contrib-python==4.6.0.66
packaging==21.3
Pillow==9.3.0
pyparsing==3.0.9
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.1
scikit-image==0.19.3
scipy==1.9.3
tifffile==2022.10.10
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.13
wandb==0.12.21
